src/score_consumer/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
rich==13.7.1
descriptastorus==2.6.1
scikit-learn==1.1.1
numpy<2
